Add feature to change status of day to fail or achieved. Add get request for current goal.
Add feature to change status of day to fail or achieved. Through heroku, the updates to the days work. However, more work must be done on what to do afterwards. Add get request for current goal, but need to figure out logic for conditional rendering of days counter or add new goal button.